[refactor] store 구조 개선 및 가이드 추가
