+++ shared/store/editorStore/useEditorStore.ts
@@ -1,10 +1,11 @@
 import { create } from 'zustand';
 import { devtools } from 'zustand/middleware';
 
+import { EditorState } from '@/shared/types/block';
+
 import { createBlockSlice } from './slices/blockSlice';
 import { createImageSlice } from './slices/imageSlice';
 import { createUISlice } from './slices/uiSlice';
-import { EditorState } from './types';
 
 export const useEditorStore = create<EditorState>()(
   devtools((...a) => ({
